SL-SL-1: moved session search related stuff to one folder
